fix: change files structure
