+++ assignments/remote-control/server/src/consts.ts
@@ -1,0 +1,11 @@
+export enum COMMANDS {
+  mouse_up = "mouse_up",
+  mouse_down = "mouse_down",
+  mouse_left = "mouse_left",
+  mouse_right = "mouse_right",
+  draw_rectangle = "draw_rectangle",
+  draw_square = "draw_square",
+  draw_circle = "draw_circle",
+  mouse_position = "mouse_position",
+  prnt_scrn = "prnt_scrn",
+}
