+++ assignments/remote-control/server/src/nut-commands.ts
@@ -1,27 +1,13 @@
 import {
-  down,
-  left,
-  mouse,
-  Point,
-  Region,
-  right,
-  screen,
-  up,
-} from "@nut-tree/nut-js";
-
-import Jimp from "jimp";
-
-export enum COMMANDS {
-  mouse_up = "mouse_up",
-  mouse_down = "mouse_down",
-  mouse_left = "mouse_left",
-  mouse_right = "mouse_right",
-  draw_rectangle = "draw_rectangle",
-  draw_square = "draw_square",
-  draw_circle = "draw_circle",
-  mouse_position = "mouse_position",
-  prnt_scrn = "prnt_scrn",
-}
+  mousePosition,
+  moveMouseDown,
+  moveMouseLeft,
+  moveMouseRight,
+  moveMouseUp,
+} from "./move-cursor.js";
+import { drawCircle, drawRectangle } from "./draw-figures.js";
+import { prntScrn } from "./screen-actions.js";
+import { COMMANDS } from "./consts.js";
 
 export const parseCommands = (
   message: string
@@ -47,79 +33,6 @@ export const parseCommands = (
   }
 
   return [foundCommand];
-};
-
-const moveMouseUp = async (...args: number[]): Promise<void> => {
-  await mouse.move(up(args[0] as number));
-};
-
-const moveMouseDown = async (...args: number[]): Promise<void> => {
-  await mouse.move(down(args[0] as number));
-};
-
-const moveMouseLeft = async (...args: number[]): Promise<void> => {
-  await mouse.move(left(args[0] as number));
-};
-
-const moveMouseRight = async (...args: number[]): Promise<void> => {
-  await mouse.move(right(args[0] as number));
-};
-
-const drawRectangle = async (...args: number[]): Promise<void> => {
-  let [width, height] = args;
-  if (!width && !height) {
-    return;
-  }
-
-  if (width && !height) {
-    height = width;
-  }
-  await mouse.drag(right(width as number));
-  await mouse.drag(down(height as number));
-  await mouse.drag(left(width as number));
-  await mouse.drag(up(height as number));
-};
-
-const drawCircle = async (...args: number[]): Promise<void> => {
-  const [radius] = args;
-  if (!radius) {
-    return;
-  }
-  const { x: centerX, y: centerY } = await mouse.getPosition();
-
-  const numPoints = 1440;
-  await moveMouseRight(radius);
-
-  const circlePath = [];
-
-  for (let i = 0; i <= numPoints; i++) {
-    const angle = i * ((2 * Math.PI) / numPoints);
-    const x = centerX + radius * Math.cos(angle);
-    const y = centerY + radius * Math.sin(angle);
-    circlePath.push(new Point(x, y));
-  }
-  await mouse.drag(circlePath);
-};
-
-const mousePosition = async (): Promise<string> => {
-  const points = await mouse.getPosition();
-  const position = `${COMMANDS.mouse_position} ${Object.values(points).join(
-    ","
-  )}`;
-
-  return position;
-};
-
-const prntScrn = async (): Promise<string> => {
-  const points = await mouse.getPosition();
-  const regionToGrab = new Region(points.x - 100, points.y - 100, 200, 200);
-  const region = await screen.grabRegion(regionToGrab);
-  const image = new Jimp(region);
-
-  const base64string = await image.getBase64Async(Jimp.MIME_PNG);
-  const base64prefixFree = base64string.replace("data:image/png;base64,", "");
-
-  return `${COMMANDS.prnt_scrn} ${base64prefixFree}`;
 };
 
 export const nutFunctions = {
